Move mapper out of the gql dir
